Import callback from utils
